--- src/main/java/com/mobble/mobbleserver/domain/articleLike/repository/ArticleLikeRepository.java
@@ -1,8 +1,0 @@
-package com.mobble.mobbleserver.domain.articleLike.repository;
-
-import com.mobble.mobbleserver.domain.articleLike.entity.ArticleLike;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ArticleLikeRepository extends JpaRepository<ArticleLike, Long> {
-
-}
