+++ src/main/java/com/mobble/mobbleserver/like/LikeType.java
@@ -1,0 +1,12 @@
+package com.mobble.mobbleserver.like;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum LikeType {
+    CLUB,
+    ARTICLE,
+    COMMENT
+}
