+++ src/main/java/com/mobble/mobbleserver/like/articleLike/dto/response/ArticleLikeToggleResponseDto.java
@@ -1,0 +1,7 @@
+package com.mobble.mobbleserver.like.articleLike.dto.response;
+
+import com.mobble.mobbleserver.like.LikeType;
+
+public record ArticleLikeToggleResponseDto(
+) {
+}
