+++ src/main/java/com/mobble/mobbleserver/like/clubLike/repository/ClubLikeRepository.java
@@ -1,6 +1,6 @@
-package com.mobble.mobbleserver.domain.clubLikes.repository;
+package com.mobble.mobbleserver.like.clubLike.repository;
 
-import com.mobble.mobbleserver.domain.clubLikes.entity.ClubLike;
+import com.mobble.mobbleserver.like.clubLike.entity.ClubLike;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ClubLikeRepository extends JpaRepository<ClubLike, Long> {
