+++ src/main/java/com/mobble/mobbleserver/like/commentLike/entity/CommentLike.java
@@ -1,6 +1,6 @@
-package com.mobble.mobbleserver.domain.articleLike.entity;
+package com.mobble.mobbleserver.like.commentLike.entity;
 
-import com.mobble.mobbleserver.domain.article.entity.Article;
+import com.mobble.mobbleserver.domain.comment.entity.Comment;
 import com.mobble.mobbleserver.domain.member.entity.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -11,30 +11,30 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ArticleLike {
+public class CommentLike {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "article_like_id")
+    @Column(name = "comment_like_id")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "article_id")
-    private Article article;
+    @JoinColumn(name = "comment_id")
+    private Comment comment;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
     private Member member;
 
     @Builder
-    private ArticleLike(Article article, Member member) {
-        this.article = article;
+    private CommentLike(Comment comment, Member member) {
+        this.comment = comment;
         this.member = member;
     }
 
-    public static ArticleLike createArticleLike(Article article, Member member) {
-        return ArticleLike.builder()
-                .article(article)
+    public static CommentLike createcommentLike(Comment comment, Member member) {
+        return CommentLike.builder()
+                .comment(comment)
                 .member(member)
                 .build();
     }
