+++ src/main/java/com/mobble/mobbleserver/like/commentLike/repository/CommentLikeRepository.java
@@ -1,6 +1,6 @@
-package com.mobble.mobbleserver.domain.commentLike.repository;
+package com.mobble.mobbleserver.like.commentLike.repository;
 
-import com.mobble.mobbleserver.domain.commentLike.entity.CommentLike;
+import com.mobble.mobbleserver.like.commentLike.entity.CommentLike;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentLikeRepository extends JpaRepository<CommentLike, Long> {
